Refs #17223. Added a new gui ui.
